update mock data
